+++ zcash_client_backend/src/data_api/wallet.rs
@@ -522,7 +522,7 @@ pub fn propose_shielding<DbT, ParamsT, InputsT, ChangeT, CommitmentTreeErrT>(
     shielding_threshold: Zatoshis,
     from_addrs: &[TransparentAddress],
     to_account: <DbT as InputSource>::AccountId,
-    min_confirmations: u32,
+    confirmations_policy: ConfirmationsPolicy,
 ) -> Result<
     Proposal<ChangeT::FeeRule, Infallible>,
     ProposeShieldingErrT<DbT, CommitmentTreeErrT, InputsT, ChangeT>,
@@ -547,7 +547,7 @@ where
             from_addrs,
             to_account,
             (chain_tip_height + 1).into(),
-            min_confirmations,
+            confirmations_policy,
         )
         .map_err(Error::from)
 }
@@ -2247,7 +2247,7 @@ pub fn shield_transparent_funds<DbT, ParamsT, InputsT, ChangeT>(
     usk: &UnifiedSpendingKey,
     from_addrs: &[TransparentAddress],
     to_account: <DbT as InputSource>::AccountId,
-    min_confirmations: u32,
+    confirmations_policy: ConfirmationsPolicy,
 ) -> Result<NonEmpty<TxId>, ShieldErrT<DbT, InputsT, ChangeT>>
 where
     ParamsT: consensus::Parameters,
@@ -2263,7 +2263,7 @@ where
         shielding_threshold,
         from_addrs,
         to_account,
-        min_confirmations,
+        confirmations_policy,
     )?;
 
     create_proposed_transactions(